finishing basic models

+++ src/main/java/com/advella/advellabackend/model/Product.java
@@ -6,7 +6,7 @@ public class Product {
     public int productId;
     public String title;
     public String detail;
-    public int moneyAmount;
+    public float moneyAmount;
     public String pickUpLocation;
     public Date postedDateTime;
     public Date deadline;
